% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pen.R
\name{pen}
\alias{pen}
\title{Penalty Function Computation}
\usage{
pen(omega, penalty, lambda, gamma = NULL)
}
\arguments{
\item{omega}{A numeric scalar or vector at which the penalty is evaluated.}

\item{penalty}{A character vector specifying one or more penalty types.
Available options include:
\enumerate{
\item "lasso": Least absolute shrinkage and selection operator
\insertCite{tibshirani1996regression,friedman2008sparse}{grasps}.
\item "atan": Arctangent type penalty \insertCite{wang2016variable}{grasps}.
\item "exp": Exponential type penalty \insertCite{wang2018variable}{grasps}.
\item "lq": Lq penalty \insertCite{frank1993statistical,fu1998penalized,fan2001variable}{grasps}.
\item "lsp": Log-sum penalty \insertCite{candes2008enhancing}{grasps}.
\item "mcp": Minimax concave penalty \insertCite{zhang2010nearly}{grasps}.
\item "scad": Smoothly clipped absolute deviation \insertCite{fan2001variable,fan2009network}{grasps}.
}}

\item{lambda}{A non-negative scalar or vector of the same length as
\code{penalty} specifying the regularization parameter.}

\item{gamma}{A scalar or vector of the same length as \code{penalty}
specifying the additional parameter for the penalty function.
The defaults are:
\enumerate{
\item "atan": 0.005
\item "exp": 0.01
\item "lq": 0.5
\item "lsp": 0.1
\item "mcp": 3
\item "scad": 3.7
}}
}
\value{
A data frame containing:
\describe{
\item{omega}{The input \code{omega} values.}
\item{penalty}{The penalty type for each row.}
\item{lambda}{The regularization parameter used.}
\item{gamma}{The additional penalty parameter used.}
\item{value}{The computed penalty value.}
}
}
\description{
Compute the penalty function.
}
\references{
\insertAllCited{}
}
