% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.adjmat.R
\name{plot.adjmat}
\alias{plot.adjmat}
\title{Plot Function for S3 Class "adjmat"}
\usage{
\method{plot}{adjmat}(x, ...)
}
\arguments{
\item{x}{An object inheriting from S3 class \code{"adjmat"}, typically
returned by \code{\link[grasps]{prec_to_adj}}.}

\item{...}{Additional arguments passed to \code{\link[ggplot2]{ggplot}}.}
}
\value{
A heatmap of class \code{ggplot} showing the matrix entries.
The plot title also reports matrix dimension and sparsity.
}
\description{
Visualize an adjacency matrix as a heatmap. This function is shared by
objects returned from \code{\link[grasps]{prec_to_adj}}.
}
\examples{
library(grasps)

## reproducibility for everything
set.seed(1234)

## block-structured precision matrix based on SBM
sim <- gen_prec_sbm(p = 30, K = 3,
                    within.prob = 0.25, between.prob = 0.05,
                    weight.dists = list("gamma", "unif"),
                    weight.paras = list(c(shape = 20, rate = 10),
                                        c(min = 0, max = 5)),
                    cond.target = 100)
## visualization
plot(sim)

## n-by-p data matrix
library(MASS)
X <- mvrnorm(n = 20, mu = rep(0, 30), Sigma = sim$Sigma)

## precision matrix: adaptive lasso; BIC
prec <- grasps(X = X, membership = sim$membership, penalty = "adapt", crit = "BIC")

## precision matrix visualization
plot(prec)

## performance
performance(hatOmega = prec$hatOmega, Omega = sim$Omega)

## adjacency matrix: diagonal = 0; raw partial correlations;
##                   no thresholding; weighted network
adj <- prec_to_adj(prec$hatOmega,
                   diag.zero = TRUE, absolute = FALSE,
                   threshold = NULL, weighted = TRUE)

## adjacency matrix visualization
plot(adj)
}
