% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparsify_block_diag.R
\name{sparsify_block_diag}
\alias{sparsify_block_diag}
\title{Groupwise Block-Diagonal Sparsifier}
\usage{
sparsify_block_diag(mat, membership, group.diag = NULL)
}
\arguments{
\item{mat}{A p-by-p precision-like matrix specifying the base matrix to be
masked.}

\item{membership}{An integer vector specifying the group membership.
The length of \code{membership} must be consistent with the dimension p.}

\item{group.diag}{(default = \code{NULL})
\enumerate{
\item \code{NULL}: Make \code{mat} block-diagonal, i.e., only keep all
within-group blocks.
\item An integer vector specifying which within-group blocks are further
reduced to diagonal-only structure (based on the block-diagonal form).
}}
}
\value{
A list containing:
\describe{
\item{Omega}{The masked precision matrix.}
\item{Sigma}{The covariance matrix, i.e., the inverse of \code{Omega}.}
\item{sparsity}{Proportion of zero entries in \code{Omega}.}
\item{membership}{An integer vector specifying the group membership.}
}
}
\description{
Make a precision-like matrix block-diagonal according to group membership,
keeping only within-group blocks. Optionally, further reduce selected
within-group blocks to diagonal-only (identity-like) structure.
}
\examples{
## reproducibility for everything
set.seed(1234)

## precision matrix estimation
X <- matrix(rnorm(200), 10, 20)
membership <- c(rep(1,5), rep(2,5), rep(3,4), rep(4,6))
est <- grasps(X, membership = membership, penalty = "lasso", crit = "BIC")

## block-diagonalizel; within-group block for group 3 made diagonal-only.
res <- sparsify_block_diag(est$hatOmega, membership, group.diag = 3)

## visualization
visualize(res$Omega, res$membership)

}
